Refactor `ContentSection` for reusability
