+++ src/components/sections/index.tsx
@@ -1,5 +1,6 @@
 import AboutHero from "./AboutHero";
+import ContentSection from "./ContentSection";
 import IconCardSection from "./IconCardSection";
 import ProjectsSection from "./ProjectsSection";
 
-export { AboutHero, IconCardSection, ProjectsSection };
+export { AboutHero, ContentSection, IconCardSection, ProjectsSection };
